Update to 4.9.0

+++ TMessagesProj/src/main/java/org/telegram/messenger/exoplayer2/source/dash/DashWrappingSegmentIndex.java
@@ -34,7 +34,7 @@ public final class DashWrappingSegmentIndex implements DashSegmentIndex {
   }
 
   @Override
-  public int getFirstSegmentNum() {
+  public long getFirstSegmentNum() {
     return 0;
   }
 
@@ -44,22 +44,23 @@ public final class DashWrappingSegmentIndex implements DashSegmentIndex {
   }
 
   @Override
-  public long getTimeUs(int segmentNum) {
-    return chunkIndex.timesUs[segmentNum];
+  public long getTimeUs(long segmentNum) {
+    return chunkIndex.timesUs[(int) segmentNum];
   }
 
   @Override
-  public long getDurationUs(int segmentNum, long periodDurationUs) {
-    return chunkIndex.durationsUs[segmentNum];
+  public long getDurationUs(long segmentNum, long periodDurationUs) {
+    return chunkIndex.durationsUs[(int) segmentNum];
   }
 
   @Override
-  public RangedUri getSegmentUrl(int segmentNum) {
-    return new RangedUri(null, chunkIndex.offsets[segmentNum], chunkIndex.sizes[segmentNum]);
+  public RangedUri getSegmentUrl(long segmentNum) {
+    return new RangedUri(
+        null, chunkIndex.offsets[(int) segmentNum], chunkIndex.sizes[(int) segmentNum]);
   }
 
   @Override
-  public int getSegmentNum(long timeUs, long periodDurationUs) {
+  public long getSegmentNum(long timeUs, long periodDurationUs) {
     return chunkIndex.getChunkIndex(timeUs);
   }
