+++ TMessagesProj/src/main/java/org/telegram/tgnet/RequestDelegateInternal.java
@@ -1,5 +1,5 @@
 package org.telegram.tgnet;
 
 public interface RequestDelegateInternal {
-    void run(int response, int errorCode, String errorText, int networkType);
+    void run(long response, int errorCode, String errorText, int networkType);
 }
